+++ tests/access/test_epo_espacenet.py
@@ -2,14 +2,16 @@
 # (c) 2022 Andreas Motl <andreas.motl@ip-tools.org>
 import pytest
 
-from patzilla.access.epo.espacenet.client import espacenet_description, espacenet_claims, espacenet_abstract
+from patzilla.access.epo.espacenet.api import espacenet_description, espacenet_claims, espacenet_abstract
 
 
 def test_espacenet_abstract_success(app_request):
     """
     Acquire "abstract" section of valid patent document from Espacenet.
+
+    TODO: Impossible to get abstract for document EP0666666B1.
     """
-    result = espacenet_abstract(document_number="EP0666666B1")
+    result = espacenet_abstract(document_number="EP0666666A2")
     assert result["source"] == "espacenet"
     assert result["lang"] == "en"
     assert "A non-quota access indicator is circulated among nodes" in result["xml"]
@@ -21,7 +23,7 @@ def test_espacenet_abstract_failure(app_request):
     """
     with pytest.raises(KeyError) as ex:
         espacenet_abstract(document_number="EP123A2")
-    assert ex.match('No section "biblio" at Espacenet for "EP123A2"')
+    assert ex.match('No section "biblio" at Espacenet for "EP0000123A2"')
 
 
 def test_espacenet_description_success(app_request):
@@ -40,7 +42,7 @@ def test_espacenet_description_failure(app_request):
     """
     with pytest.raises(KeyError) as ex:
         espacenet_description(document_number="EP123A2")
-    assert ex.match('No section "description" at Espacenet for "EP123A2"')
+    assert ex.match('No section "description" at Espacenet for "EP0000123A2"')
 
 
 def test_espacenet_claims_success(app_request):
@@ -59,4 +61,4 @@ def test_espacenet_claims_failure(app_request):
     """
     with pytest.raises(KeyError) as ex:
         espacenet_claims(document_number="EP123A2")
-    assert ex.match('No section "claims" at Espacenet for "EP123A2"')
+    assert ex.match('No section "claims" at Espacenet for "EP0000123A2"')